Melhoria no tratamento do arquivo

+++ back-end/app/Http/Controllers/Api/ObjetoController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Models\Objeto;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\File;
 
 class ObjetoController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      *
@@ -24,19 +26,29 @@ class ObjetoController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
+
     public function store(Request $request)
     {
-       
-        $objeto = Objeto::create($request->except(['file']));
 
-        if($request->file('file')->isValid()){
-
-            $path = $request->file('file')->store('objeto/'.$objeto->id);
-            $objeto->path=$path;
-            $objeto->save();
-        }
+        $objeto = Objeto::create($request->except('file'));
+        $objeto->filename=hash_file('md5', $request->file);
+        $objeto->size = $request->file->getSize();
+        $objeto->extension = $request->file->getClientOriginalExtension();
+        $objeto->path = $request->file->storeAs('objeto/' . $objeto->id, $objeto->filename.'.'. $objeto->extension);
+        $objeto->save();
         
+        
+    }
 
+    public function update(Request $request, Objeto $objeto)
+    {
+        $objeto->nome=$request->nome;
+        $objeto->descricao=$request->descricao;
+        $objeto->filename=hash_file('md5', $request->file);
+        $objeto->size = $request->file->getSize();
+        $objeto->extension = $request->file->getClientOriginalExtension();
+        $objeto->path = $request->file->storeAs('objeto/' . $objeto->id, $objeto->filename.'.'. $objeto->extension);
+        $objeto->update();
     }
 
     /**
@@ -57,10 +69,7 @@ class ObjetoController extends Controller
      * @param  \App\Models\Objeto  $objeto
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Objeto $objeto)
-    {
-        $objeto->update($request->all());
-    }
+
 
     /**
      * Remove the specified resource from storage.
@@ -70,6 +79,16 @@ class ObjetoController extends Controller
      */
     public function destroy(Objeto $objeto)
     {
-        $objeto->delete();
+        if (File::deleteDirectory(storage_path('app/files/objeto/') . $objeto->id))
+            $objeto->delete();
+        else return "Não apagou!";
+    }
+
+    public function download(Objeto $objeto)
+    {
+        // Check if file exists in app/storage/file folder
+
+        $file_path = storage_path('app/files/') . $objeto->path;
+        return response()->download($file_path);
     }
 }
